fix: fix duplicate module reoslution problems
