--- application/backend/src/cli/__init__.py
@@ -1,2 +1,0 @@
-# Copyright (C) 2025 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
